fix: Change type of schemaVersion to string
This makes FDSNStationXML>schemaVersion compatible with schema>version;
right now one can specify any semantic version patch, build, or release
info in the schema, but cannot use that in a StationXML file due to the
type. This change makes the two fields consistent and allows for
semantic-like version names to propagate to the data file.