feat: Make StageGain optional element of Stage
This change makes StageGain consistent with all other elements of Stage
by making it optional. It also allows for the use case of a Stage
containing only a number and a unique resourceID, which can then refer
to an already defined or external entity (like a RESP file) that
contains a partial response.